requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
websockets==12.0
backoff==2.2.1
python-multipart==0.0.6